Made changes requested by dmisharo
